requirements.txt
attrs==18.2.0
autobahn==18.12.1
Automat==0.7.0
constantly==15.1.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
mypy==0.650
mypy-extensions==0.4.1
numpy==1.15.4
opencv-python==3.4.5.20
PyHamcrest==1.9.0
pyzmq==17.1.2
six==1.12.0
Twisted==18.9.0
txaio==18.8.1
typed-ast==1.1.1
websockets==7.0
zope.interface==4.6.0
